Fixed replyingToUserId type to Long in CommentRequest

+++ src/main/java/com/wsu/crimsonconnect/dto/CommentRequest.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class CommentRequest {
     private String content;
-    private Integer replyingToUserId;
+    private Long replyingToUserId;
 }
